+++ src/api/tests/schemas/test_utils.py
@@ -323,7 +323,7 @@ def test_save_statiques_number_of_database_queries(db_session):
             1,  # list already existing pdc
             6 * n_statiques,  # look for existing PDC-related entries
             6,  # create non existing PDC-related entries
-            2,  # update foreign keys
+            6,  # update foreign keys
             6 * n_statiques,  # PDC to statique
             2 * n_statiques,  # get station's linked operational unit and link it
         )